requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
django-allauth==0.57.0
django-filter==23.3
django-compressor==4.4